Create Audio Movement v1 (Gradual Movement)
This uses for loops to control the position of the jaw servo. The best angles to work with are 10 to 50 degrees (found using ServoRead sketch). The Lion jaw opens at the start of the audio, but it does not ever work again. It stays open, even when lines 125 to 127 are completed (as seen on Monitor).